WIP: Add support for second segment
The second segment has to be registered before the primary
segment is created as the QEMU and ACPI code currently has
hardcoded assumptions about the PCI hole.

Also the second segment should like outside the currently
mapped memory.

The segment is not yet reported in the MCFG.

The segment is now visible in in the qom tree but not inside
the OS.

QEMU is also built in debug mode and tracing enabled.
Note: Compiling in debug mode needs some of the gcc flags to be
turned off. This is not true in maintline qemu.

We need to examine why our configuration triggers these errors.

> info qtree

  dev: pci-virt-host, id ""
    MCFG = 274877906944 (0x4000000000)
    mcfg_size = 1048576 (0x100000)
    bus: 1.pcie.0
      type PCIE
      dev: pci-virt-device, id ""
        addr = 00.0
        romfile = ""
        rombar = 1 (0x1)
        multifunction = false
        command_serr_enable = true
        x-pcie-lnksta-dllla = true
        x-pcie-extcap-init = true
        class Host bridge, addr 00:00.0, pci id 1b36:0008 (sub
1af4:1100)

Signed-off-by: Manohar Castelino <manohar.r.castelino@intel.com>