WIP: DEBUG: Enable support for kernel_irqchip=off and split
As we do not need the IOAPIC anymore we should be able to support
both split mode and the off mode.

In split mode we have no PIC and IOAPIC for virt and that works.
In off mode we move the APIC modeling into QEMU. This does not work
even though QEMU is trying to inject the interrupt into the APIC.
The kernel never sees them. This needs to be debugged.

Signed-off-by: Manohar Castelino <manohar.r.castelino@intel.com>